modif place Intent dans ScannerCodeBarreActivity

+++ app/src/main/java/com/example/myfridge/AjoutAliment.java
@@ -56,7 +56,7 @@ public class AjoutAliment extends AppCompatActivity {
 
     public void AjouterAlimentBDD(View view) {
         EditText editQuantite =findViewById(R.id.editTextQuantite);
-        quantite= Integer.parseInt(editQuantite.getText().toString());
+        quantite = (int) Integer.parseInt(editQuantite.getText().toString());
         //Ajout de l'aliment dans la base de données
         AlimentsOperations alimentsOperations = new AlimentsOperations(this);
         alimentsOperations.open();
